refactor: rename sirene and siret as it was before
